removed api calls and replaced with sdk calls. Other cleanups.
